Rename domhtml.ts to htmltables.ts
A pure data module named like widthtables.ts, its sibling in kind: the
tag, interface, reflection and event-handler tables dom.ts builds its
element classes from.
